drift-monitoring/requirements.txt
# Kafka clients
confluent-kafka>=2.0.0
kafka-python>=2.0.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Online learning for adaptor
river>=0.18.0

# Visualization
matplotlib>=3.7.0

